backend/requirements.txt
cassandra-driver
tornado
pytest
pytest-mock
